+++ lib/puree/xml_extractor/research_output.rb
@@ -32,16 +32,17 @@ module Puree
       end
 
       # Digital Object Identifier (first one, if many)
-      # @return [String, nil]
+      # @return [Puree::Model::DOI, nil]
       def doi
         multiple_dois = dois
         multiple_dois.empty? ? nil : multiple_dois.first
       end
 
       # Digital Object Identifiers
-      # @return [Array<String>]
+      # @return [Array<Puree::Model::DOI>]
       def dois
-        xpath_query_for_multi_value '/electronicVersions/electronicVersion[@type="wsElectronicVersionDoiAssociation"]/doi'
+        xpath_result = xpath_query_for_multi_value '/electronicVersions/electronicVersion[@type="wsElectronicVersionDoiAssociation"]/doi'
+        xpath_result.map { |i| Puree::Model::DOI.new i }
       end
 
       # @return [Array<Puree::Model::File>]