+++ test/xml_extractor/xml_extractor_research_output_test.rb
@@ -30,8 +30,8 @@ class TestXMLExtractorResearchOutput < Minitest::Test
     assert_instance_of String, x.description
     refute_empty x.description
 
-    assert_instance_of String, x.doi
-    refute_empty x.doi
+    assert_instance_of Puree::Model::DOI, x.doi
+    refute_empty x.doi.to_s
 
     assert_instance_of Array, x.files
     assert_instance_of Puree::Model::File, x.files.first