+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 2.5.0 2018-11-23
+## 2.5.0 2018-11-26
+### Added
+Research output, dataset - URL-friendly dois.
+
 ### Changed
 - For Pure API 511 due to change in source XML for Scopus ID in research output.